requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.2.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.4.4
oauthlib==3.3.1
packaging==26.1
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
urllib3==2.6.3
